Finish exercise 4.21

+++ controllers/blogs.js
@@ -30,6 +30,29 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET) // request.token is set in tokenExtractor middleware, raises JsonWebTokenError or TokenExpiredError if invalid
+  if (!decodedToken.id) { // decodedToken should contain username and id fields
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  if (!user) {
+    return response.status(404).json({ error: 'user not found' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (!blog.user) {
+    return response.status(404).json({ error: 'no user associated with this blog' }) // should not be possible
+  }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(403).json({ error: 'user did not create this blog' }) // forbidden
+  }
+
   const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
   if (deletedBlog) {
     response.status(204).end() // No content
